Add a non-relational database (mongodb) & Docker

+++ src/main/java/com/example/springboot/interfaces/UserFinder.java
@@ -2,6 +2,8 @@ package com.example.springboot.interfaces;
 
 import com.example.springboot.entities.User;
 
+import java.math.BigInteger;
+
 public interface UserFinder {
-    User getUserDetails(int id);
+    User getUserDetails(BigInteger id);
 }
